fix: resolve type mismatch and format issues (#731)
- consolidate() expects f32 not f64 — remove incorrect cast
- Server dream config: unwrap_or_default() for Option<usize> returns 0
  (wrong), use unwrap_or(DreamConfig::default().*) instead
- Format: restructure server dream config block for rustfmt compliance

+++ crates/uteke-core/src/dream.rs
@@ -302,7 +302,8 @@ impl crate::Uteke {
 
     fn phase_dedup(&self, namespace: Option<&str>, dry_run: bool) -> Result<PhaseResult, Error> {
         // Threshold from config (#731) — very similar; tune down for more aggressive merges.
-        let result = self.consolidate(namespace, self.dream_config.dedup_threshold as f64, dry_run)?;
+        let result =
+            self.consolidate(namespace, self.dream_config.dedup_threshold, dry_run)?;
         let summary = if result.merged == 0 {
             format!(
                 "✓ {} duplicate pairs found, 0 merged{}",
